@iv-menshenin 2024-12-07
```
goos: linux
goarch: amd64
pkg: github.com/iv-menshenin/lyceum/collection
cpu: Intel(R) Core(TM) i7-9700F CPU @ 3.00GHz
                           │ CollectionGet_10mln │  CollectionPushPop  │  CollectionInsert   │
                           │       sec/op        │    sec/op     vs base   │    sec/op     vs base   │
*/Last-8                            1.659n ±  5%
*/First-8                           1.696n ±  6%
*/Mid-8                             1.642n ±  2%
*/Forward-8                         2.182n ± 10%
*/Backward-8                        2.353n ±  4%
*/Random-8                          2.992n ±  4%
*/Push_All-8                                       19.71n ± 11%
*/Push_Len_All-8                                   19.25n ± 12%
*/Push_Then_Pop-8                                  19.35n ± 10%
*/Push_Push_Then_Pop-8                             41.55n ± 15%
*/Push_Push_Then_Pop_Len-8                         43.58n ± 11%
*/Pop_All-8                                        11.79n ±  7%
*/Push_1K-8                                                              20.15µ ± 18%
*/Push_65K-8                                                             1.372m ± 10%
*/Push_1M-8                                                              21.29m ± 11%
*/Push_10M-8                                                             209.1m ± 12%
geomean                             2.035n         23.22n        ? ¹ ²   3.331m        ? ¹ ²
¹ benchmark set differs from baseline; geomeans may not be comparable
² ratios must be >0 to compute geomean

                           │ CollectionGet_10mln │  CollectionPushPop  │  CollectionInsert  │
                           │      allocs/op      │ allocs/op   vs base │  allocs/op   vs base   │
*/Last-8                            0.000 ± 0%
*/First-8                           0.000 ± 0%
*/Mid-8                             0.000 ± 0%
*/Forward-8                         0.000 ± 0%
*/Backward-8                        0.000 ± 0%
*/Random-8                          0.000 ± 0%
*/Push_All-8                                       0.000 ± 0%
*/Push_Len_All-8                                   0.000 ± 0%
*/Push_Then_Pop-8                                  0.000 ± 0%
*/Push_Push_Then_Pop-8                             0.000 ± 0%
*/Push_Push_Then_Pop_Len-8                         0.000 ± 0%
*/Pop_All-8                                        0.000 ± 0%
*/Push_1K-8                                                               1.000 ± 0%
*/Push_65K-8                                                              128.0 ± 0%
*/Push_1M-8                                                              1.953k ± 0%
*/Push_10M-8                                                             19.54k ± 0%
geomean                                        ¹               ? ² ¹ ³    264.4       ? ² ³
¹ summaries must be >0 to compute geomean
² benchmark set differs from baseline; geomeans may not be comparable
³ ratios must be >0 to compute geomean
```